check file exist before deleting; change channel to private (#9334)

+++ app/Gateway/AxisMigs/Gateway.php
@@ -916,7 +916,10 @@ class Gateway extends Base\Gateway
             {
                 fclose($this->curlLog);
 
-                unlink($this->curlLogPath);
+                if (file_exists($this->curlLogPath) === true)
+                {
+                    unlink($this->curlLogPath);
+                }
             }
         }
