+++ app/Gateway/FirstData/Gateway.php
@@ -971,7 +971,10 @@ class Gateway extends Base\Gateway
             {
                 fclose($this->curlLog);
 
-                unlink($this->curlLogPath);
+                if (file_exists($this->curlLogPath) === true)
+                {
+                    unlink($this->curlLogPath);
+                }
             }
         }
